refactor: Centralise constants and split large leaderboards modules
- Add leaderboards/constants.py holding all package constants (renamed for
  clarity); remove paths.py and rename utils.py -> records.py.
- Split result_processing.py into record_fields/model_metadata/score_extraction
  and core_models.py into osai/model_sources; keep orchestrators lean.
- Reuse euroeval: import log_once and __version__, derive language-group codes
  from euroeval Language objects.
- Consolidate the URL skip-list into models_without_urls_cache.yaml; add an
  interactive pCloud retry on backup OSError; drop dead code, unused vars,
  and redundant aliases/pass-throughs; add missing docstring Args.
