+++ src/leaderboards/models_without_urls_cache.yaml
@@ -6,5 +6,20 @@ alexandrainst/Meta-Llama-3.1-8B-laerebogen-and-filtered-dialog-k2-da: true
 alexandrainst/Meta-Llama-3.1-8B-laerebogen-two-epochs: true
 danish-foundation-models/Meta-Llama-3.1-8B-laerebogen: true
 danish-foundation-models/encoder-medium-v1: false
+danish-foundation-models/munin-7b-core-it: false
+danish-foundation-models/munin-7b-core-pt: false
+danish-foundation-models/munin-7b-core-pt-2: false
+danish-foundation-models/munin-7b-core-pt-3: false
+danish-foundation-models/munin-7b-open-it: false
+danish-foundation-models/munin-7b-open-pt: false
+danish-foundation-models/munin-7b-v0.1dev0: false
+fresh-electra-small: false
+fresh-xlm-roberta-base: false
+mhenrichsen/danskgpt-chat-v2.1: false
 qwen3.5-397b: false
+skole-gpt-mixtral: false
+syvai/danskgpt-chat-llama3-70b: false
+syvai/llama3-da-base: false
 test_model: true
+xai/grok-3-beta: false
+xai/grok-3-mini-beta: false
