+++ src/leaderboards/queue_env.py
@@ -236,11 +236,14 @@ def acquire_single_instance_lock(lock_path: Path) -> int:
             f"different path if you need to run a second instance."
         )
         sys.exit(1)
+
     # Truncate the lock file first so stale bytes from a longer old PID are
     # removed before writing the current PID.
     os.ftruncate(fd, 0)
     os.write(fd, f"{os.getpid()}\n".encode())
+
     # Flush file contents to disk so other processes can reliably read the PID
     # associated with the lock holder.
     os.fsync(fd)
+
     return fd
